Broke up types.ts into folder
