+++ src/components/items/ItemEditorPanel.vue
@@ -1,6 +1,6 @@
 <!-- Right-side drawer panel for creating or editing an item's details, schedule, color, and workspace memberships. -->
 <script setup lang="ts">
-import { areItemsEqual, constructEmptyItem, generateItemId, type Item } from '@/types'
+import { areItemsEqual, constructEmptyItem, generateItemId, type Item } from '@/types/items'
 import { computed, nextTick, ref, useTemplateRef, watch } from 'vue'
 import { useItemsStore } from '@/stores/items'
 import { useWorkspacesStore } from '@/stores/workspaces'
