+++ src/components/roadmap/roadmap-utils.ts
@@ -1,5 +1,5 @@
 // Roadmap type definitions, date-range and interval computation helpers, and the xForDate coordinate conversion function.
-import type { Item } from '@/types'
+import type { Item } from '@/types/items'
 import { MSEC_IN_DAY, type DateRange } from '@/utils/dates'
 
 // === TYPES ===
