+++ src/db/index.ts
@@ -1,6 +1,7 @@
 // IndexedDB persistence layer: opens and caches the Chronicle database and exposes CRUD helpers for items, workspaces, and settings.
 import { openDB, type DBSchema, type IDBPDatabase } from 'idb'
-import type { Item, Workspace } from '@/types'
+import type { Item } from '@/types/items'
+import type { Workspace } from '@/types/workspaces'
 import { toRaw } from 'vue'
 import type { RoadmapInterval } from '@/components/roadmap/roadmap-utils'
 import {
