+++ src/stores/items/initialization.ts
@@ -1,6 +1,6 @@
 // Loads all persisted items from IndexedDB into the items store on startup.
 import { getAllItems } from '@/db'
-import type { Item } from '@/types'
+import type { Item } from '@/types/items'
 import type { Ref } from 'vue'
 
 export function useItemsInitialization(items: Ref<Record<string, Item>>) {
