+++ src/stores/workspaces/workspace-membership.ts
@@ -1,6 +1,6 @@
 // Provides moveItem, addItemToWorkspace, and removeItemFromWorkspace operations for managing workspace item membership.
 import { putWorkspace } from '@/db'
-import type { Workspace } from '@/types'
+import type { Workspace } from '@/types/workspaces'
 import type { Ref } from 'vue'
 import { useItemsStore } from '../items'
